increase the datatype for scores from int to long long this fixes https://bugs.eressea.de/view.php?id=2136

+++ src/json.c
@@ -94,7 +94,7 @@ int json_export(stream * out, int flags) {
             cJSON_AddItemToObject(json, itoa36(f->no), data = cJSON_CreateObject());
             cJSON_AddStringToObject(data, "name", f->name);
             cJSON_AddStringToObject(data, "email", f->email);
-            cJSON_AddNumberToObject(data, "score", f->score);
+            cJSON_AddNumberToObject(data, "score", (double)f->score);
         }
     }
     if (flags) {
